docs: dokka

+++ src/test/kotlin/Testdata.kt
@@ -1,6 +1,6 @@
 data class Testdata(
-    var matrix: Array<DoubleArray>,
-    var solvingVector: DoubleArray,
+    var matrix: Array<IntArray>,
+    var solvingVector: IntArray,
     var resultVector: DoubleArray?,
     var inverse: Array<DoubleArray>?
 )
